+++ Doctrine/Phpcr/AbstractBlock.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Cmf\Bundle\BlockBundle\Doctrine\Phpcr;
+
+use Symfony\Cmf\Bundle\BlockBundle\Model\AbstractBlock as AbstractBlockModel;
+
+abstract class AbstractBlock extends AbstractBlockModel
+{
+
+}