+++ Admin/StringBlockAdmin.php
@@ -28,7 +28,7 @@ class StringBlockAdmin extends Admin
             ->with('form.group_general')
             ->add('parentDocument', 'doctrine_phpcr_odm_tree', array('root_node' => $this->contentRoot, 'choice_list' => array(), 'select_root_node' => true))
             ->add('name', 'text')
-            ->add('content', 'textarea')
+            ->add('body', 'textarea')
             ->end()
         ;
     }