+++ Tests/Functional/Block/SimpleBlockServiceTest.php
@@ -4,7 +4,7 @@ namespace Symfony\Cmf\Bundle\BlockBundle\Tests\Functional\Block;
 
 use Sonata\BlockBundle\Block\BlockContext;
 use Symfony\Cmf\Bundle\BlockBundle\Block\SimpleBlockService,
-    Symfony\Cmf\Bundle\BlockBundle\Document\SimpleBlock;
+    Symfony\Cmf\Bundle\BlockBundle\Doctrine\Phpcr\SimpleBlock;
 
 class SimpleBlockServiceTest extends \PHPUnit_Framework_TestCase
 {